Extend make targets to all Grafana Cloud dashboards
Discover dashboards using `wildcard` function instead of hardcoding
filenames in the Makefile.

The function produces file names that include the directory, e.g.
`dashboards/foo.jsonnet` which means the directory becomes part of the
stem so it has to be created as part of the recipe for the target
instead of statically. This actually simplifies things and allows us to
get rid of two rules.